==> Scripts/33_dram_product_from_picr_day11.R <==
## ---------------------------
##
## Script: 23_dram_product_from_picrust_day11
##
## Purpose: Create T/F df for picrust predicted genomes from day11 based on DRAM polymer rules
##
## Author: Ikaia Leleiwi
##
## Date Created: 2022-03-28
##
## Copyright (c) Ikaia Leleiwi, 2022
## Email: ileleiwi@gmail.com
##
## ---------------------------
##
## Notes: requires GH_in_picrust_day11.tsv and DRAM rules sheets
##        apply DRAM SOM rules for function presence absence on dbcan_picrust
##        using function heatmap form
##        must include an oligo cleaving and a backbone cleaving for TRUE
##   
##
## ---------------------------

## set working directory

setwd(paste0("/Users/ikaialeleiwi/Desktop/Lab/Salmonella_NIH/Lactobacillus/",
             "Omics/Metagenome/rRNA/16S_From_All_Bins/MQHQ_Picrust2/",
             "MQHQ_Picrust2/"))

## ---------------------------

library(tidyverse)

##Read in data

#DRAM genome summary form
genome_form_cazy <- read_tsv("Data/genome_summary_form.tsv") %>%
  filter(header == "CAZY") 

#DRAM function heatmap form
function_heatmap_form_cazy <- read_tsv("Data/function_heatmap_form.tsv") %>%
  filter(category == "CAZy") %>%
  select(-gene_symbol)

#dbCAN ids in picrust day11
dbcan_picrust <- read_tsv("Clean_Data/GH_in_picrust_day11.tsv") 


####Functions####
checkID <- function(row, function_df, id){
  #check for id in function_id column - T if id is in cell, F if id not in cell
  cell <- function_df %>%
    slice(row) %>%
    pull(function_ids)
  
  return(str_detect(cell, id))
}


makeFunctBinDf <- function(columns, bin_row){
  #Make function/bin df
  df <- data.frame(row.names = bin_row)
  for(i in columns){
    c_add <- data.frame(i = "x")
    df <- cbind(df, c_add)
  }
  colnames(df) <- columns
  return(df)
}


functionBool <- function(sub_names, column, function_df, dbcan){
  #Make function boolean for bin
  ids_bin <- dbcan[dbcan[,column] > 0, c("target_name", column)]
  df_out <- makeFunctBinDf(columns = sub_names, bin_row = column)
  
  for(i in sub_names){
    funct <- function_df %>%
      filter(str_detect(function_name, i))
    if(nrow(funct) == 1){
      detected <- 0
      for(j in ids_bin$target_name){
        detected <- detected + checkID(row = 1, function_df = funct, id=j)
      }
      df_out[column, i] <- detected > 0
    }
    if(nrow(funct) == 2){
      detected1 <- 0
      detected2 <- 0
      for(j in ids_bin$target_name){
        detected1 <- detected1 + checkID(row = 1, function_df = funct, id=j)
        detected2 <- detected2 + checkID(row = 2, function_df = funct, id=j)
      }
      df_out[column, i] <- detected1 > 0 & detected2 > 0
    }
  }
  
  return(df_out)
}



buildProduct <- function(sub_names, function_df, dbcan){
  #create df with rownames = bins, colnames = function, TRUE/FALSE if present
  columns <- colnames(dbcan)[-1]
  df <- functionBool(sub_names,
                     column = columns[1],
                     function_df,
                     dbcan)
  #progressbar
  n_iter <- length(columns) - 1
  pb <- txtProgressBar(min = 0,
                       max = n_iter,
                       style = 3,
                       width = n_iter, 
                       char = "=") 
  
  init <- numeric(n_iter)
  end <- numeric(n_iter)
  
  for(i in  2:length(columns)){
    init[i] <- Sys.time()
    
    #code
    df <- rbind(df, functionBool(sub_names,
                                 column = columns[i],
                                 function_df,
                                 dbcan))
    
    end[i] <- Sys.time()
    
    #progress bar
    setTxtProgressBar(pb, i)
    time <- round(seconds_to_period(sum(end - init)), 0)
    est <- n_iter * (mean(end[end != 0] - init[init != 0])) - time
    remaining <- round(seconds_to_period(est), 0)
    cat(paste(" // Execution time:", time,
              " // Estimated time remaining:", remaining), "")
  }
  return(df)
}

#check for polymer type (clevage, oligo + backbone, or oligo or backbone) 
#needs both oligo and backbone clevage or just 1

#get clean list of polymers
substrate_names <- function_heatmap_form_cazy %>%
  pull(function_name) %>%
  str_remove(" \\(.+\\)") %>%
  str_remove(" Cleavage| cleavage") %>%
  str_remove(" Oligo| oligo") %>%
  str_remove(" Backbone| backbone") %>%
  unique()

dram_product <- buildProduct(substrate_names,
                             function_heatmap_form_cazy,
                             dbcan_picrust)

dram_product <- dram_product %>%
  rownames_to_column(var = "bins")

write_tsv(dram_product, "Clean_Data/dram_product_picrust_day11.tsv")
